requirements.txt
cohere
gunicorn
flask
python-dotenv
requests